fix: guard GoogleOAuthProvider against missing VITE_GOOGLE_CLIENT_ID in CI
Passing clientId="" to GoogleOAuthProvider crashed the React tree in
CI (no env var set), preventing the auth form from rendering and
breaking the E2E smoke test.

Fix:
- main.tsx: only wrap with GoogleOAuthProvider when clientId is non-empty
- Auth.tsx: move useGoogleLogin into a dedicated GoogleLoginButton
  component; only render OAuth section when VITE_GOOGLE_CLIENT_ID is set
- GoogleLoginButton renders nothing when provider is not in the tree
