+++ src/pages/Auth.tsx
@@ -1,26 +1,12 @@
 import { useState } from "react";
 import { useNavigate, useLocation } from "react-router-dom";
 import { Brain, ArrowRight, Loader2 } from "lucide-react";
-import { useGoogleLogin } from "@react-oauth/google";
 import { authService } from "../services/auth.service";
 import { useAuthStore } from "../stores/auth.store";
 import { toast } from "sonner";
+import { GoogleLoginButton } from "../components/auth/GoogleLoginButton";
 
-// Add new providers here — UI updates automatically
-const OAUTH_PROVIDERS = [
-  {
-    key: "google",
-    label: "Continue with Google",
-    icon: (
-      <svg className="w-5 h-5" viewBox="0 0 24 24" aria-hidden="true">
-        <path fill="#4285F4" d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"/>
-        <path fill="#34A853" d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"/>
-        <path fill="#FBBC05" d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l3.66-2.84z"/>
-        <path fill="#EA4335" d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"/>
-      </svg>
-    ),
-  },
-];
+const GOOGLE_CLIENT_ID = import.meta.env.VITE_GOOGLE_CLIENT_ID as string | undefined;
 
 export default function AuthPage() {
   const [isLogin, setIsLogin] = useState(true);
@@ -47,12 +33,6 @@ export default function AuthPage() {
       setOauthLoading(null);
     }
   };
-
-  const googleLogin = useGoogleLogin({
-    onSuccess: (response) => handleOAuthSuccess("google", response.access_token),
-    onError: () => toast.error("Google login failed. Please try again."),
-    flow: "implicit",
-  });
 
   const [formData, setFormData] = useState({
     email: "",
@@ -216,41 +196,27 @@ export default function AuthPage() {
             </div>
           </form>
 
-          {/* OAuth providers */}
-          <div className="mt-6">
-            <div className="relative">
-              <div className="absolute inset-0 flex items-center">
-                <div className="w-full border-t border-gray-700" />
+          {/* OAuth providers — only shown when at least one provider is configured */}
+          {GOOGLE_CLIENT_ID && (
+            <div className="mt-6">
+              <div className="relative">
+                <div className="absolute inset-0 flex items-center">
+                  <div className="w-full border-t border-gray-700" />
+                </div>
+                <div className="relative flex justify-center text-sm">
+                  <span className="px-2 bg-gray-900/50 text-gray-400">Or continue with</span>
+                </div>
               </div>
-              <div className="relative flex justify-center text-sm">
-                <span className="px-2 bg-gray-900/50 text-gray-400">Or continue with</span>
-              </div>
-            </div>
 
-            <div className="mt-4 space-y-3">
-              {OAUTH_PROVIDERS.map((p) => {
-                const isProviderLoading = oauthLoading === p.key;
-                const handleClick =
-                  p.key === "google" ? () => googleLogin() : undefined;
-                return (
-                  <button
-                    key={p.key}
-                    type="button"
-                    onClick={handleClick}
-                    disabled={!!oauthLoading || isLoading}
-                    className="w-full flex items-center justify-center gap-3 py-2.5 px-4 border border-gray-700 rounded-lg bg-gray-900/60 text-gray-200 text-sm font-medium hover:bg-gray-800 transition-colors disabled:opacity-50 disabled:cursor-not-allowed"
-                  >
-                    {isProviderLoading ? (
-                      <Loader2 className="animate-spin" size={18} />
-                    ) : (
-                      p.icon
-                    )}
-                    {p.label}
-                  </button>
-                );
-              })}
+              <div className="mt-4 space-y-3">
+                <GoogleLoginButton
+                  isLoading={oauthLoading === "google"}
+                  onSuccess={(token) => handleOAuthSuccess("google", token)}
+                />
+                {/* Add more provider buttons here as needed */}
+              </div>
             </div>
-          </div>
+          )}
         </div>
       </div>
     </main>
